+++ android/src/main/java/com/dayaki/reactnativehealthkits/HealthKitsPackage.kt
@@ -1,12 +1,12 @@
 package com.dayaki.reactnativehealthkits
 
-import com.facebook.react.TurboReactPackage
+import com.facebook.react.BaseReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.module.model.ReactModuleInfo
 import com.facebook.react.module.model.ReactModuleInfoProvider
 
-class HealthKitsPackage : TurboReactPackage() {
+class HealthKitsPackage : BaseReactPackage() {
     override fun getModule(name: String, reactContext: ReactApplicationContext): NativeModule? {
         return if (name == HealthKitsModule.NAME) {
             HealthKitsModule(reactContext)
@@ -23,7 +23,6 @@ class HealthKitsPackage : TurboReactPackage() {
                 HealthKitsModule.NAME,
                 false,  // canOverrideExistingModule
                 false,  // needsEagerInit
-                true,   // hasConstants
                 false,  // isCxxModule
                 true    // isTurboModule
             )